requirements.txt
# requirements.txt

# Django
Django==4.2.7
djangorestframework==3.14.0
django-extensions==3.2.3
django-environ==0.11.2
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Supabase
supabase
gotrue==1.3.0
postgrest==0.13.2

# JWT handling  
PyJWT==2.8.0
cryptography==41.0.7

# HTTP requests
requests==2.31.0
httpx

# Task queue (for future use)
celery==5.3.4
redis==5.0.1

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring and logging
sentry-sdk[django]==1.38.0

# Development tools
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Type checking
mypy==1.7.1
django-stubs==4.2.7